+++ integrationTests/consensus/consensus_test.go
@@ -51,10 +51,10 @@ func initNodesAndTest(numNodes, consensusSize, numInvalid uint32, roundTime uint
 				fmt.Println("process block invalid ", header.GetRound(), header.GetNonce(), getPkEncoded(nodes[i].pk))
 				return process.ErrBlockHashDoesNotMatch
 			}
-			nodes[i].blkProcessor.CreateBlockHeaderCalled = func(body data.BodyHandler, round uint32, haveTime func() bool) (handler data.HeaderHandler, e error) {
+			nodes[i].blkProcessor.CreateBlockHeaderCalled = func(body data.BodyHandler, round uint64, haveTime func() bool) (handler data.HeaderHandler, e error) {
 				return nil, process.ErrAccountStateDirty
 			}
-			nodes[i].blkProcessor.CreateBlockCalled = func(round uint32, haveTime func() bool) (handler data.BodyHandler, e error) {
+			nodes[i].blkProcessor.CreateBlockCalled = func(round uint64, haveTime func() bool) (handler data.BodyHandler, e error) {
 				return nil, process.ErrWrongTypeAssertion
 			}
 		}
@@ -63,7 +63,7 @@ func initNodesAndTest(numNodes, consensusSize, numInvalid uint32, roundTime uint
 	return nodes, advertiser, concMap
 }
 
-func startNodesWithCommitBlock(nodes []*testNode, mutex *sync.Mutex, nonceForRoundMap map[uint32]uint64, totalCalled *int) error {
+func startNodesWithCommitBlock(nodes []*testNode, mutex *sync.Mutex, nonceForRoundMap map[uint64]uint64, totalCalled *int) error {
 	for _, n := range nodes {
 		n.blkProcessor.CommitBlockCalled = func(blockChain data.ChainHandler, header data.HeaderHandler, body data.BodyHandler) error {
 			n.blkProcessor.NrCommitBlockCalled++
@@ -85,13 +85,13 @@ func startNodesWithCommitBlock(nodes []*testNode, mutex *sync.Mutex, nonceForRou
 	return nil
 }
 
-func checkBlockProposedEveryRound(numCommBlock uint32, nonceForRoundMap map[uint32]uint64, mutex *sync.Mutex, chDone chan bool, t *testing.T) {
+func checkBlockProposedEveryRound(numCommBlock uint64, nonceForRoundMap map[uint64]uint64, mutex *sync.Mutex, chDone chan bool, t *testing.T) {
 	for {
 		mutex.Lock()
 
-		minRound := ^uint32(0)
-		maxRound := uint32(0)
-		if uint32(len(nonceForRoundMap)) >= numCommBlock {
+		minRound := ^uint64(0)
+		maxRound := uint64(0)
+		if uint64(len(nonceForRoundMap)) >= numCommBlock {
 			for k := range nonceForRoundMap {
 				if k > maxRound {
 					maxRound = k
@@ -127,7 +127,7 @@ func runFullConsensusTest(t *testing.T, consensusType string) {
 	consensusSize := uint32(4)
 	numInvalid := uint32(0)
 	roundTime := uint64(4000)
-	numCommBlock := uint32(10)
+	numCommBlock := uint64(10)
 	nodes, advertiser, _ := initNodesAndTest(numNodes, consensusSize, numInvalid, roundTime, consensusType)
 
 	mutex := &sync.Mutex{}
@@ -142,7 +142,7 @@ func runFullConsensusTest(t *testing.T, consensusType string) {
 	fmt.Println("Start consensus...")
 	time.Sleep(time.Second * 1)
 
-	nonceForRoundMap := make(map[uint32]uint64)
+	nonceForRoundMap := make(map[uint64]uint64)
 	totalCalled := 0
 	err := startNodesWithCommitBlock(nodes, mutex, nonceForRoundMap, &totalCalled)
 	assert.Nil(t, err)
@@ -150,7 +150,7 @@ func runFullConsensusTest(t *testing.T, consensusType string) {
 	chDone := make(chan bool, 0)
 	go checkBlockProposedEveryRound(numCommBlock, nonceForRoundMap, mutex, chDone, t)
 
-	extraTime := uint32(2)
+	extraTime := uint64(2)
 	endTime := time.Duration(roundTime) * time.Duration(numCommBlock+extraTime) * time.Millisecond
 	select {
 	case <-chDone:
@@ -198,7 +198,7 @@ func runConsensusWithNotEnoughValidators(t *testing.T, consensusType string) {
 	fmt.Println("Start consensus...")
 	time.Sleep(time.Second * 1)
 
-	nonceForRoundMap := make(map[uint32]uint64)
+	nonceForRoundMap := make(map[uint64]uint64)
 	totalCalled := 0
 	err := startNodesWithCommitBlock(nodes, mutex, nonceForRoundMap, &totalCalled)
 	assert.Nil(t, err)
