+++ integrationTests/mock/txProcessorMock.go
@@ -8,12 +8,12 @@ import (
 )
 
 type TxProcessorMock struct {
-	ProcessTransactionCalled         func(transaction *transaction.Transaction, round uint32) error
+	ProcessTransactionCalled         func(transaction *transaction.Transaction, round uint64) error
 	SetBalancesToTrieCalled          func(accBalance map[string]*big.Int) (rootHash []byte, err error)
 	ProcessSmartContractResultCalled func(scr *smartContractResult.SmartContractResult) error
 }
 
-func (etm *TxProcessorMock) ProcessTransaction(transaction *transaction.Transaction, round uint32) error {
+func (etm *TxProcessorMock) ProcessTransaction(transaction *transaction.Transaction, round uint64) error {
 	return etm.ProcessTransactionCalled(transaction, round)
 }
