+++ node/node.go
@@ -100,7 +100,7 @@ type Node struct {
 	isMetachainActive        bool
 	txStorageSize            uint32
 	currentSendingGoRoutines int32
-	bootstrapRoundIndex      uint32
+	bootstrapRoundIndex      uint64
 }
 
 // ApplyOptions can set up different configurable options of a Node instance
